Saad 20.11.2022 (12:38 am)
The establishment of a new improved files directory/structure was made
where all(most) of the code files will be saved in the website folder
containing static and template folders. The main folder which
the application is run by is still a stand-alone file. New code was added
mainly making basic blueprints and routing for the login
logout and sign up pages, along with other additives.